requirements.txt
flask
python-dotenv
Pillow
google-generativeai
pdf2image
pytesseract
pdfplumber
